delete the auth table from the database and simplification of the merchant table properties according to MLD needs
