+++ backend/src/Merchant/merchant.entity.ts
@@ -1,6 +1,5 @@
 import { UUID } from 'crypto';
 import { Entity, PrimaryGeneratedColumn, Column, OneToOne, JoinColumn } from 'typeorm';
-import { Auth } from '../Auth/auth.entity';
 
 @Entity('merchant')
 export class Merchant {
@@ -9,10 +8,7 @@ export class Merchant {
     id: UUID;
 
     @Column({ length: 100 })
-    lastName: string;
-
-    @Column({ length: 100 })
-    firstName: string;
+    siret: string;
 
     @Column({ length: 100 })
     email: string;
@@ -20,17 +16,5 @@ export class Merchant {
     @Column({ length: 1000 })
     password: string;
 
-    @Column({ length: 1000 })
-    passwordValidation: string;
-
-    @Column({ length: 100 })
-    siret: string;
-
-    @Column({ length: 100 })
-    headQuarter: string;
-
-    @OneToOne(type => Auth)
-    @JoinColumn()
-    auth: Auth;
 
 }
