+++ backend/src/Merchant/merchant.module.ts
@@ -3,10 +3,10 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { Merchant } from './merchant.entity';
 import { MerchantController } from './merchant.controller';
 import { MerchantService } from './merchant.service';
-import { Auth } from '../Auth/auth.entity';
+//import { Auth } from '../Auth/auth.entity';
 
 @Module({
-    imports: [TypeOrmModule.forFeature([Merchant, Auth])],
+    imports: [TypeOrmModule.forFeature([Merchant])],
     controllers: [MerchantController],
     providers: [MerchantService],
     exports: [MerchantService],
