+++ frontend/src/routes/forms/login/+page.svelte
@@ -8,7 +8,7 @@
 
 	const handleLogin = async () => {
 		try {
-			const response = await fetch(`http://localhost:3000/auth/login`, {
+			const response = await fetch(`http://localhost:3000/login`, {
 				method: 'POST',
 				headers: {
 					'Content-Type': 'application/json'
